requirements.txt
click==8.1.3
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
plotly==5.13.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
tenacity==8.2.0
Werkzeug==2.2.2
zipp==3.12.1